+++ packages/solverpy/src/solverpy/solver/smt/__init__.py
@@ -14,6 +14,7 @@ class solverpy.solver.smt.bitwuzla.Bitwuzla extends solverpy.solver.shellsolver.
 class solverpy.solver.smt.llm2smt.Llm2smt extends solverpy.solver.shellsolver.ShellSolver
 class solverpy.solver.smt.primo.Primo extends solverpy.solver.shellsolver.ShellSolver
 class solverpy.solver.smt.opensmt.Opensmt extends solverpy.solver.shellsolver.ShellSolver
+class solverpy.solver.smt.yices.Yices extends solverpy.solver.shellsolver.ShellSolver
 
 ```
 
@@ -25,6 +26,7 @@ class solverpy.solver.smt.opensmt.Opensmt extends solverpy.solver.shellsolver.Sh
 | [`Llm2smt`][solverpy.solver.smt.llm2smt.Llm2smt] | `llm2smt` | QF_EUF solver |
 | [`Primo`][solverpy.solver.smt.primo.Primo] | `primo` | QF_LRA and QF_UF solver |
 | [`Opensmt`][solverpy.solver.smt.opensmt.Opensmt] | `opensmt` | SMT solver with interpolation and proofs |
+| [`Yices`][solverpy.solver.smt.yices.Yices] | `yices-smt2` | Yices 2; SMT-LIB2 front end of `yices` |
 
 """
 
@@ -33,6 +35,7 @@ from .cvc5 import Cvc5
 from .llm2smt import Llm2smt
 from .opensmt import Opensmt
 from .primo import Primo
+from .yices import Yices
 from .z3 import Z3
 
-__all__ = ["Bitwuzla", "Cvc5", "Llm2smt", "Opensmt", "Primo", "Z3"]
+__all__ = ["Bitwuzla", "Cvc5", "Llm2smt", "Opensmt", "Primo", "Yices", "Z3"]