+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Change Log
 
-## Unreleased changes (2026-08-10)
+## Unreleased changes (2026-08-17)
 
-* test(smt): cover the opensmt plugin [[details](https://github.com/cbboyan/solverpy/commit/) | [browse](https://github.com/cbboyan/solverpy/tree/)]
+* feat(smt): add yices solver plugin [[details](https://github.com/cbboyan/solverpy/commit/) | [browse](https://github.com/cbboyan/solverpy/tree/)]
+* test(smt): cover the opensmt plugin [[details](https://github.com/cbboyan/solverpy/commit/1637c53) | [browse](https://github.com/cbboyan/solverpy/tree/1637c53)]
 * feat(smt): add opensmt solver plugin [[details](https://github.com/cbboyan/solverpy/commit/81c47a2) | [browse](https://github.com/cbboyan/solverpy/tree/81c47a2)]
 * feat(smt): add primo solver plugin [[details](https://github.com/cbboyan/solverpy/commit/e7ca1e8) | [browse](https://github.com/cbboyan/solverpy/tree/e7ca1e8)]